Implement a Java JAR Decompiler
This is where Natty jumps the shark and goes from being a silly little side project to being something that's actually testing out some legitimately interesting new ideas in the LLM-coding space.

This decompiler (based on "Fernflower") will be used to decompile the "Header JARs" that Bazel's internal "Turbine" generates for each `java_library()` target so that I can get a stripped down representation of JUST the signatures that the LLM should know about for its dependencies, so I can avoid feeding the full implementation source file of all dependencies to the LLM. This should reduce the context necessary for the LLM to avoid it getting confused. And it should enable me to potentially enable the Natty to pass more of the dep subtree beneath a given component to the LLM to avoid the transitive dependency issue where omitting the sources of a dependency's dependency causes the LLM to be forced to hallucinate signatures of methods on types returned from its dependencies.

E.g. a dep structure like `Foo` -> `Bar` -> `Baz` where `Bar` has a func like `def passThroughDep(...) -> Baz.Something`. From `Bar`'s perspective, this makes perfect sense, because it's seen the impl of `Baz` it knows everything about wtv that `Baz.Something` name refers to. But then from `Foo` 's perspective, it only sees `Bar`'s impl, but not `Baz`'s (by design - I could of course expand the scope so components see all sources of everything, but then you may as well not use Natty, cuz you're just feeding your entire codebase to the LLM at a certain point).