Implement `_natty_header()` Rule & Call From `natty_java_library()` Macro
This now declares a target to generate the "Natty Header" for every LLM-generated Natty Java component.

For now, it's not yet wired through to the dependents as I still need to add some additional context to these header files in the form of LLM-generated usage instructions so that the LLM stands a chance at understanding certain unfortunately stripped down things like `void update(long var1){}` where `var1` should be interpreted as millis, for example.

Once I generate this additional short context to the header we can try wiring this through in LLM calls instead of the full source code, to try reducing the noise in what the LLM uses to do codegen.